Popup system now working
The Popup System:
Popup is a static class which provides a way of showing a popup to the player where the player can choose something or can only confirm.

Usage: (Decission)
the popup system supports the decission option functionality of the inventory (used for choosing "throw", "use", etc.
The workflow for the Popup is the flollowing:
- Call Popup.NewDecission and provide some parameters if you want
- you can add additional decission by calling Popup.AddDecissionOption
- transition into the Popup by calling Popup.Show()

Therere is also a new TextConfirmation Popup. Which will display a text and can be closed with the "Use" key. (or canceled with the cancel key if enabled)

Something important: Popup privides a Property called Popup.Showing which can be used to see if the Popup is shown. (example: moving in inventory while popup is there) so there might need to be some additions for that.

Also: There is now a Property in GameManager called "MenuOpen" which returns true if any Menu is open (anvil, inventory, grimoire, skills/technic,...); can also be used

Finally: Added a new Input for InventoryChoice4 so we can currently have 4 different choices in a Popup Decission