Finalized generation upgrades
Introducing the new technology of "Floors" which now store the data for a floor and allow the parallel genereation (theoretically) to accelerate start up time.

Fixed the map being read from the wrong direction.

There is now a universal global RNG which should be preferably used (can also be used outside the main thread